report_app/requirements.txt
reportlab
psycopg2-binary
matplotlib